requirements.txt
numpy
torch
imageio
imgviz
matplotlib
opencv_python
pyqt5
pyyaml
tqdm
tensorboard
